Tidy
- Undo changes in ff_port.c
  - Keep default baudrate 9600 for serial ports
  - Remove spurious PRINT(), there's already a DEBUG() with that info. Please PRINT() in your application.. :-)
- Change field names in UBX_NAV_TIMELS_V0_GROUP0_t as per Interface Description:
  - iTOW, not iTow
  - reserved0, not __pad0
  - reserved1, not __pad1
- Added UBX_NAV_TIMELS_V0_SRCOFCURRLS_... and UBX_NAV_TIMELS_V0_SRCOFCURRLSCHANGE_... fields for the magic numbers in those message fields
- Makefile recipes for the library
  - Add release flags to build (compile and link)
  - Add LIBPREFIX variable to choose libraray target directory, defaults to /usr/local
- Changed all types in EPOCH_t to fundamental types (int)
- Added warning on the leap second event stuff in epoch.
- Move new section in the README.md a but further down, update text slightly
- Convert tabs to spaces
- Move dev-test/* to contrib/lib-test/

+++ ff/ff_epoch.h
@@ -278,16 +278,18 @@ typedef struct EPOCH_s
     int                 sigCnoHistNav[EPOCH_SIGCNOHIST_NUM];
 
     bool                haveLeapSeconds;
-    int8_t              leapSeconds;
+    int                 leapSeconds;
 
+    // Information as per UBX-NAV-TIMELS
+    // Warning: This is subject to changed (simplified) in future versions of this library. Users are advised
+    //          to directly use the UBX-NAV-TIMELS message if they have to rely on u-blox specific behaviour.
     bool                haveLeapSecondEvent;
-    uint8_t             srcOfCurrLs;
-    uint8_t             srcOfLsChange;
-    int8_t              lsChange;
-    int32_t             timeToLsEvent;
-    uint16_t            dateOfLsGpsWn;
-    uint16_t            dateOfLsGpsDn;
-
+    int                 srcOfCurrLs;
+    int                 srcOfLsChange;
+    int                 lsChange;
+    int                 timeToLsEvent;
+    int                 dateOfLsGpsWn;
+    int                 dateOfLsGpsDn;
 
     // Private
     int          _haveFix;
